pyproject.toml
[project]
name = "gepa-mcp"
version = "0.1.0"
description = "MCP server integrating GEPA (Genetic-Evolutionary Prompt Architecture) for automatic prompt optimization"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.9.2",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gepa_mcp"]
